Remove old system. Stand up tests on new system
