--- components/lark-codegen/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "lark-codegen"
-version = "0.1.0"
-authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-lark-mir = { path = "../lark-mir" }
-tempfile = "2.0"
-cc = "1.0"
