--- components/lark-codegen2/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "lark-codegen2"
-version = "0.1.0"
-authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-lark-mir2 = { path = "../lark-mir2" }
-lark-hir = { path = "../lark-hir" }
-lark-type-check = { path = "../lark-type-check" }
-lark-query-system = { path = "../lark-query-system" }
-lark-entity = { path = "../lark-entity" }
-lark-string = { path = "../lark-string" }
-lark-ty = { path = "../lark-ty" }
-lark-error = { path = "../lark-error" }
-parser = { path = "../parser" }
-tempfile = "2.0"
-intern = { path = "../intern" }
-ast = { path = "../ast" }
-cc = "1.0"
