+++ components/lark-eval/Cargo.toml
@@ -6,3 +6,13 @@ edition = "2018"
 
 [dependencies]
 lark-mir = { path = "../lark-mir" }
+lark-hir = { path = "../lark-hir" }
+lark-type-check = { path = "../lark-type-check" }
+lark-query-system = { path = "../lark-query-system" }
+lark-entity = { path = "../lark-entity" }
+lark-string = { path = "../lark-string" }
+lark-ty = { path = "../lark-ty" }
+lark-error = { path = "../lark-error" }
+parser = { path = "../parser" }
+intern = { path = "../intern" }
+ast = { path = "../ast" }
