+++ components/lark-eval/src/lib.rs
@@ -1,12 +1,19 @@
+use ast::AstDatabase;
+use intern::{Intern, Untern};
+use lark_entity::{EntityData, ItemKind, LangItem};
 use lark_mir::{
-    BinOp, Context, DefId, Definition, Function, Operand, Place, Rvalue, Statement, StatementKind,
+    BasicBlock, FnBytecode, MirDatabase, Operand, OperandData, Place, PlaceData, Rvalue,
+    RvalueData, Statement, StatementKind, Variable,
 };
+use lark_query_system::LarkDatabase;
+use parser::ReaderDatabase;
 use std::collections::HashMap;
 use std::fmt;
 
 #[derive(Clone, Debug)]
 pub enum Value {
     Void,
+    Bool(bool),
     I32(i32),
     Str(String),
     Struct(HashMap<String, Value>),
@@ -21,22 +28,12 @@ impl fmt::Display for Value {
             match self {
                 Value::I32(i) => i.to_string(),
                 Value::Str(s) => s.clone(),
+                Value::Bool(b) => b.to_string(),
                 Value::Reference(r) => format!("reference to {}", r),
                 Value::Void => "<void>".into(),
                 Value::Struct(s) => format!("{:?}", s),
             }
         )
-    }
-}
-
-#[derive(Debug)]
-pub struct CallFrame {
-    locals: Vec<Value>,
-}
-
-impl CallFrame {
-    fn new() -> CallFrame {
-        CallFrame { locals: vec![] }
     }
 }
 
@@ -65,147 +62,204 @@ impl IOHandler {
     }
 }
 
-pub fn eval_operand(frame: &mut CallFrame, operand: &Operand) -> Value {
-    match operand {
-        Operand::ConstantInt(i) => Value::I32(*i),
-        Operand::ConstantString(s) => Value::Str(s.clone()),
-        Operand::Move(m) => match m {
-            Place::Local(source_var_id) => frame.locals[*source_var_id].clone(),
-            Place::Static(_) => unimplemented!("Moving from static data not currently supported"),
-            Place::Field(source_var_id, field_name) => match &frame.locals[*source_var_id] {
-                Value::Struct(s) => s[field_name].clone(),
-                _ => unimplemented!("Field access of non-struct value"),
-            },
+pub fn eval_place(
+    db: &mut LarkDatabase,
+    fn_bytecode: &FnBytecode,
+    place: Place,
+    variables: &mut HashMap<Variable, Vec<Value>>,
+) -> Value {
+    let place_data = &fn_bytecode.tables[place];
+
+    match place_data {
+        PlaceData::Entity(entity) => match entity.untern(db) {
+            EntityData::LangItem(LangItem::True) => Value::Bool(true),
+            EntityData::LangItem(LangItem::False) => Value::Bool(false),
+            _ => unimplemented!("EntityData not yet support in eval"),
         },
-        Operand::Copy(m) => match m {
-            Place::Local(source_var_id) => frame.locals[*source_var_id].clone(),
-            Place::Static(_) => unimplemented!("Moving from static data not currently supported"),
-            Place::Field(source_var_id, field_name) => match &frame.locals[*source_var_id] {
-                Value::Struct(s) => s[field_name].clone(),
-                _ => unimplemented!("Field access of non-struct value"),
-            },
-        },
+        PlaceData::Variable(variable) => {
+            let stack = variables.get(variable).unwrap();
+            stack.last().unwrap().clone()
+        }
+        x => unimplemented!("PlaceData not yet support in eval: {:#?}", x),
     }
 }
 
-fn eval_rvalue(
-    context: &Context,
-    frame: &mut CallFrame,
-    rvalue: &Rvalue,
+pub fn eval_operand(
+    db: &mut LarkDatabase,
+    fn_bytecode: &FnBytecode,
+    operand: Operand,
+    variables: &mut HashMap<Variable, Vec<Value>>,
+) -> Value {
+    let operand_data = &fn_bytecode.tables[operand];
+
+    match operand_data {
+        OperandData::Copy(place) | OperandData::Move(place) => {
+            eval_place(db, fn_bytecode, *place, variables)
+        }
+        _ => unimplemented!("Operand not yet supported"),
+    }
+}
+
+pub fn eval_rvalue(
+    db: &mut LarkDatabase,
+    fn_bytecode: &FnBytecode,
+    rvalue: Rvalue,
+    variables: &mut HashMap<Variable, Vec<Value>>,
     io_handler: &mut IOHandler,
 ) -> Value {
-    match rvalue {
-        Rvalue::Use(ref operand) => eval_operand(frame, operand),
-        Rvalue::Call(def_id, args) => {
-            match &context.definitions[*def_id] {
-                Definition::Fn(f) => {
-                    let mut new_frame = CallFrame::new();
-                    new_frame.locals.push(Value::Void); // return value
-                    for arg in args {
-                        new_frame.locals.push(eval_operand(frame, arg));
-                    }
-                    let num_temps = f.local_decls.len() - 1 - f.arg_count;
-                    for _ in 0..num_temps {
-                        new_frame.locals.push(Value::Void);
-                    }
-                    eval_fn(context, &mut new_frame, f, io_handler);
-                    let result = new_frame.locals[0].clone();
-                    result
-                }
-                Definition::Struct(ref s) => {
-                    let mut new_obj = HashMap::new();
-                    for i in 0..s.fields.len() {
-                        new_obj.insert(s.fields[i].name.clone(), eval_operand(frame, &args[i]));
-                    }
-                    Value::Struct(new_obj)
-                }
-                _ => unimplemented!("Unsupported call of non-function"),
-            }
-        }
-        Rvalue::BinaryOp(bin_op, lhs_var_id, rhs_var_id) => {
-            let lhs = &frame.locals[*lhs_var_id];
-            let rhs = &frame.locals[*rhs_var_id];
+    let rvalue_data = &fn_bytecode.tables[rvalue];
 
-            match bin_op {
-                BinOp::Add => match (lhs, rhs) {
-                    (Value::I32(lhs_i32), Value::I32(rhs_i32)) => Value::I32(lhs_i32 + rhs_i32),
-                    _ => unimplemented!("Unsupported add of non-integers"),
-                },
-                BinOp::Sub => match (lhs, rhs) {
-                    (Value::I32(lhs_i32), Value::I32(rhs_i32)) => Value::I32(lhs_i32 - rhs_i32),
-                    _ => unimplemented!("Unsupported add of non-integers"),
-                },
-            }
-        }
-    }
-}
+    match rvalue_data {
+        RvalueData::Use(operand) => eval_operand(db, fn_bytecode, *operand, variables),
+        RvalueData::Call(entity, operands) => match entity.untern(db) {
+            EntityData::LangItem(LangItem::Debug) => {
+                for operand in operands.iter(fn_bytecode) {
+                    let result = eval_operand(db, fn_bytecode, operand, variables);
+                    io_handler.println(format!("{}", result));
+                }
 
-pub fn eval_stmt(
-    context: &Context,
-    frame: &mut CallFrame,
-    stmt: &Statement,
-    io_handler: &mut IOHandler,
-) {
-    match &stmt.kind {
-        StatementKind::Assign(place, rvalue) => {
-            let rval = eval_rvalue(context, frame, rvalue, io_handler);
-            match place {
-                Place::Local(target_var_id) => frame.locals[*target_var_id] = rval,
-                Place::Static(_) => unimplemented!("Assigning into static currently not supported"),
-                Place::Field(source_var_id, field_name) => {
-                    match &mut frame.locals[*source_var_id] {
-                        Value::Struct(s) => {
-                            let _ = s.insert(field_name.clone(), rval);
-                        }
-                        _ => unimplemented!("Field access of non-struct value"),
-                    }
-                }
+                Value::Void
             }
-        }
-        StatementKind::DebugPrint(place) => match place {
-            Place::Local(var_id) => {
-                io_handler.println(format!("{}", frame.locals[*var_id]));
-            }
-            Place::Static(_) => unimplemented!("Debug print of value other than local variable"),
-            Place::Field(source_var_id, field_name) => match &frame.locals[*source_var_id] {
-                Value::Struct(s) => {
-                    io_handler.println(format!("{}", s[field_name]));
+            EntityData::ItemName { .. } => {
+                let bytecode = db.fn_bytecode(*entity).value;
+
+                for (arg, param) in operands
+                    .iter(fn_bytecode)
+                    .zip(bytecode.arguments.iter(&bytecode))
+                {
+                    let arg_value = eval_operand(db, fn_bytecode, arg, variables);
+                    create_variable(variables, param);
+                    assign_to_variable(variables, param, arg_value);
                 }
-                _ => unimplemented!("Field access of non-struct value"),
-            },
+
+                let return_value = eval_function(db, &bytecode, variables, io_handler);
+
+                for argument in fn_bytecode.arguments.iter(&fn_bytecode) {
+                    pop_variable(variables, argument);
+                }
+
+                return_value
+            }
+            _ => unimplemented!("EntityData not yet supported in eval"),
         },
+        _ => unimplemented!("Rvalue not yet supported in eval"),
     }
 }
 
-pub fn eval_fn(
-    context: &Context,
-    frame: &mut CallFrame,
-    fun: &Function,
-    io_handler: &mut IOHandler,
+pub fn create_variable(variables: &mut HashMap<Variable, Vec<Value>>, variable: Variable) {
+    let variable_stack = variables.entry(variable).or_insert(Vec::new());
+    variable_stack.push(Value::Void);
+}
+
+pub fn pop_variable(variables: &mut HashMap<Variable, Vec<Value>>, variable: Variable) {
+    let variable_stack = variables.get_mut(&variable).unwrap();
+    variable_stack.pop();
+}
+
+pub fn assign_to_variable(
+    variables: &mut HashMap<Variable, Vec<Value>>,
+    variable: Variable,
+    value: Value,
 ) {
-    for block in &fun.basic_blocks {
-        for stmt in &block.statements {
-            eval_stmt(context, frame, stmt, io_handler);
+    let variable_stack = variables.get_mut(&variable).unwrap();
+    *variable_stack.last_mut().unwrap() = value;
+}
+
+pub fn eval_statement(
+    db: &mut LarkDatabase,
+    fn_bytecode: &FnBytecode,
+    statement: Statement,
+    variables: &mut HashMap<Variable, Vec<Value>>,
+    io_handler: &mut IOHandler,
+) -> Value {
+    let statement_data = &fn_bytecode.tables[statement];
+
+    match &statement_data.kind {
+        StatementKind::Expression(rvalue) => {
+            eval_rvalue(db, fn_bytecode, *rvalue, variables, io_handler)
+        }
+        StatementKind::Assign(place, rvalue) => {
+            let rhs = eval_rvalue(db, fn_bytecode, *rvalue, variables, io_handler);
+            match &fn_bytecode.tables[*place] {
+                PlaceData::Variable(variable) => assign_to_variable(variables, *variable, rhs),
+                _ => unimplemented!("PlaceData not yet supported in eval"),
+            }
+            Value::Void
+        }
+        _ => {
+            // If we get here, something went wrong. Statements like StorageLive/StorageDead
+            // Need to get handled where they won't effect the return value of the block
+            panic!("Unexpected StatementKind");
         }
     }
 }
 
-/// Starting place for eval of a project or file.
-///   * context - the full set of definitions from MIR
-///   * starting_fn - the "main" function to begin evaluation
-///   * io_handler - a helper to handle output to console or capture for testing
-pub fn eval_context(context: &Context, starting_fn: DefId, io_handler: &mut IOHandler) {
-    match context.definitions[starting_fn] {
-        Definition::Fn(ref f) => {
-            let mut frame = CallFrame::new();
-            let num_temps = f.local_decls.len() - 1 - f.arg_count;
-            frame.locals.push(Value::Void); // return value
-            for _ in 0..num_temps {
-                frame.locals.push(Value::Void);
-            }
+pub fn eval_basic_block(
+    db: &mut LarkDatabase,
+    fn_bytecode: &FnBytecode,
+    basic_block: BasicBlock,
+    variables: &mut HashMap<Variable, Vec<Value>>,
+    io_handler: &mut IOHandler,
+) -> Value {
+    let basic_block_data = &fn_bytecode.tables[basic_block];
+    let mut return_value = Value::Void;
 
-            eval_fn(context, &mut frame, f, io_handler);
+    for statement in basic_block_data.statements.iter(&fn_bytecode) {
+        let statement_data = &fn_bytecode.tables[statement];
+
+        match &statement_data.kind {
+            StatementKind::StorageLive(variable) => {
+                create_variable(variables, *variable);
+            }
+            StatementKind::StorageDead(variable) => {
+                pop_variable(variables, *variable);
+            }
+            _ => return_value = eval_statement(db, fn_bytecode, statement, variables, io_handler),
         }
-        _ => unimplemented!("Starting function is not a function definition"),
+    }
+
+    return_value
+}
+
+pub fn eval_function(
+    db: &mut LarkDatabase,
+    fn_bytecode: &FnBytecode,
+    variables: &mut HashMap<Variable, Vec<Value>>,
+    io_handler: &mut IOHandler,
+) -> Value {
+    let mut return_value = Value::Void;
+    for basic_block in fn_bytecode.basic_blocks.iter(fn_bytecode) {
+        return_value = eval_basic_block(db, fn_bytecode, basic_block, variables, io_handler);
+    }
+
+    return_value
+}
+
+pub fn eval(db: &mut LarkDatabase, io_handler: &mut IOHandler) {
+    let input_files = db.paths();
+    //let mut errors: Vec<Diagnostic> = vec![];
+
+    let mut variables: HashMap<Variable, Vec<Value>> = HashMap::new();
+    let main_name = "main".intern(&db);
+
+    for &input_file in &*input_files {
+        let entities = db.items_in_file(input_file);
+
+        for &entity in &*entities {
+            match entity.untern(&db) {
+                EntityData::ItemName {
+                    kind: ItemKind::Function,
+                    id,
+                    ..
+                } => {
+                    if id == main_name {
+                        let bytecode = db.fn_bytecode(entity);
+
+                        eval_function(db, &bytecode.value, &mut variables, io_handler);
+                    }
+                }
+                _ => {}
+            }
+        }
     }
 }
