+++ components/lark-mir/src/lib.rs
@@ -1,150 +1,170 @@
-pub type DefId = usize;
-pub type VarId = usize;
+#![feature(crate_visibility_modifier)]
+#![feature(const_fn)]
+#![feature(const_let)]
+#![feature(decl_macro)]
+#![feature(in_band_lifetimes)]
+#![feature(macro_at_most_once_rep)]
+#![feature(specialization)]
 
-// Dummy for now
-#[derive(Copy, Clone, Debug)]
-pub struct Ty {
-    def_id: DefId,
-}
+use indices::{IndexVec, U32Index};
+use lark_debug_derive::DebugWith;
+use lark_entity::Entity;
+use lark_error::WithError;
+use lark_string::global::GlobalIdentifier;
+use lark_ty::declaration::DeclarationTables;
+use lark_type_check as typecheck;
+use parser::pos::{HasSpan, Span, Spanned};
+use std::sync::Arc;
 
-#[derive(Debug)]
-pub struct SourceInfo;
+mod fn_bytecode;
 
-//Lark MIR representation of a single function
-#[derive(Debug)]
-pub struct Function {
-    pub basic_blocks: Vec<BasicBlock>,
-
-    //First local = return value pointer
-    //Followed by arg_count parameters to the function
-    //Followed by user defined variables and temporaries
-    pub local_decls: Vec<LocalDecl>,
-
-    pub arg_count: usize,
-
-    pub name: String,
-}
-
-impl Function {
-    pub fn new(return_ty: Ty, mut args: Vec<LocalDecl>, name: String) -> Function {
-        let arg_count = args.len();
-        let mut local_decls = vec![LocalDecl::new_return_place(return_ty)];
-        local_decls.append(&mut args);
-
-        Function {
-            basic_blocks: vec![],
-            local_decls,
-            arg_count,
-            name,
+salsa::query_group! {
+    pub trait MirDatabase: typecheck::TypeCheckDatabase + AsRef<DeclarationTables> {
+        fn fn_bytecode(key: Entity) -> WithError<Arc<FnBytecode>> {
+            type FnBytecodeQuery;
+            use fn fn_bytecode::fn_bytecode;
         }
     }
-
-    pub fn new_temp(&mut self, ty: Ty) -> VarId {
-        self.local_decls.push(LocalDecl::new_temp(ty));
-        self.local_decls.len() - 1
-    }
-
-    pub fn push_block(&mut self, block: BasicBlock) {
-        self.basic_blocks.push(block);
-    }
 }
 
-#[derive(Debug)]
-pub struct Struct {
-    pub fields: Vec<Field>,
-    pub name: String,
+/*
+indices::index_type! {
+    pub struct Identifier { .. }
 }
 
-impl Struct {
-    pub fn field(mut self, name: String, ty: Ty) -> Self {
-        self.fields.push(Field { ty, name });
+#[derive(Copy, Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub struct IdentifierData {
+    pub text: GlobalIdentifier,
+}
+
+indices::index_type! {
+    pub struct Variable { .. }
+}
+
+#[derive(Copy, Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub struct VariableData {
+    pub name: Identifier,
+}
+
+*/
+
+indices::index_type! {
+    pub struct Error { .. }
+}
+
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub enum ErrorData {
+    Misc,
+    Unimplemented,
+    UnknownIdentifier { text: GlobalIdentifier },
+}
+
+/// All the data for a fn-body is stored in these tables.a
+#[derive(Clone, Debug, DebugWith, Default, PartialEq, Eq, Hash)]
+pub struct FnBytecodeTables {
+    /// Map each statement to its associated data.
+    pub statements: IndexVec<Statement, Spanned<StatementData>>,
+
+    /// The blocks that make up the code of the function
+    pub basic_blocks: IndexVec<BasicBlock, Spanned<BasicBlockData>>,
+
+    /// Map each place index to its associated data.
+    pub places: IndexVec<Place, Spanned<PlaceData>>,
+
+    /// Map each variable index to its associated data.
+    pub variables: IndexVec<Variable, Spanned<VariableData>>,
+
+    /// Map each identifier index to its associated data.
+    pub identifiers: IndexVec<Identifier, Spanned<IdentifierData>>,
+
+    /// Map each rvalue index to its associated data.
+    pub rvalues: IndexVec<Rvalue, Spanned<RvalueData>>,
+
+    /// Map each operand index to its associated data.
+    pub operands: IndexVec<Operand, Spanned<OperandData>>,
+
+    /// The data values for any `List<I>` values that appear elsewhere
+    /// in the HIR; the way this works is that all of the list value
+    /// are concatenated into one big vector, and each list just pulls
+    /// out a slice of that. Note that this just contains `u32` values
+    /// -- the actual `List<I>` remembers the index type `I` for its
+    /// own values and does the casting back and forth.
+    pub list_entries: Vec<u32>,
+}
+
+impl AsMut<FnBytecodeTables> for FnBytecodeTables {
+    fn as_mut(&mut self) -> &mut Self {
         self
     }
-
-    pub fn new(name: String) -> Self {
-        Struct {
-            name,
-            fields: vec![],
-        }
-    }
 }
 
-#[derive(Debug)]
-pub struct Field {
-    pub ty: Ty,
-    pub name: String,
+//Lark MIR representation of a single function
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub struct FnBytecode {
+    /// List of arguments to the function. The type of each argument
+    /// is given by the function signature (which can be separately queried).
+    pub arguments: List<Variable>,
+
+    /// The code of the function body, split into basic blocks
+    pub basic_blocks: List<BasicBlock>,
+
+    pub tables: FnBytecodeTables,
 }
 
-#[derive(Debug)]
-pub struct BasicBlock {
-    pub statements: Vec<Statement>,
-    pub terminator: Option<Terminator>,
+indices::index_type! {
+    pub struct BasicBlock { .. }
 }
 
-impl BasicBlock {
-    pub fn new() -> BasicBlock {
-        BasicBlock {
-            statements: vec![],
-            terminator: None,
-        }
-    }
-
-    pub fn push_stmt(&mut self, kind: StatementKind) {
-        self.statements.push(Statement {
-            source_info: SourceInfo,
-            kind,
-        });
-    }
-
-    pub fn terminate(&mut self, terminator_kind: TerminatorKind) {
-        self.terminator = Some(Terminator {
-            source_info: SourceInfo,
-            kind: terminator_kind,
-        });
-    }
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub struct BasicBlockData {
+    pub statements: List<Statement>,
+    pub terminator: Terminator,
 }
 
-#[derive(Debug)]
-pub struct Statement {
-    pub source_info: SourceInfo,
+indices::index_type! {
+    pub struct Statement { .. }
+}
+
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub struct StatementData {
     pub kind: StatementKind,
 }
 
-#[derive(Debug)]
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
 pub enum StatementKind {
     Assign(Place, Rvalue),
-    DebugPrint(Place),
+
+    /// Start a live range for the storage of the variable.
+    StorageLive(Variable),
+
+    /// End the current live range for the storage of the variable.
+    StorageDead(Variable),
+
+    Expression(Rvalue),
 }
 
-#[derive(Debug)]
-pub struct Terminator {
-    pub source_info: SourceInfo,
-    pub kind: TerminatorKind,
-}
-
-#[derive(Debug)]
-pub enum TerminatorKind {
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub enum Terminator {
     Return,
+    PassThrough,
 }
 
-#[derive(Debug)]
-pub enum Place {
-    Local(VarId),
-    Static(DefId),
-    //FIXME: this is a simplifed projection for now
-    Field(VarId, String),
+indices::index_type! {
+    pub struct Rvalue { .. }
 }
-
-#[derive(Debug)]
-pub enum Rvalue {
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub enum RvalueData {
     Use(Operand),
-    BinaryOp(BinOp, VarId, VarId),
-    //FIXME: MIR has this as a Terminator, presumably because stack can unwind
-    Call(DefId, Vec<Operand>),
+    BinaryOp(BinOp, Variable, Variable),
+    //FIXME: MIR has this as a TerminatorData, presumably because stack can unwind
+    Call(Entity, List<Operand>),
 }
 
-#[derive(Debug)]
-pub enum Operand {
+indices::index_type! {
+    pub struct Operand { .. }
+}
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub enum OperandData {
     Copy(Place),
     Move(Place),
     //FIXME: Move to Box<Constant>
@@ -152,76 +172,289 @@ pub enum Operand {
     ConstantString(String),
 }
 
-#[derive(Debug)]
+#[derive(Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
 pub enum BinOp {
     Add,
     Sub,
 }
 
-#[derive(Debug)]
-pub struct LocalDecl {
-    pub ty: Ty,
-    pub name: Option<String>,
+/// Trait implemented by the various kinds of indices that reach into
+/// the HIR; allows us to grab the vector that they correspond to.
+pub trait MirIndex: U32Index + Into<MetaIndex> {
+    type Data: Clone;
+
+    fn index_vec(mir: &FnBytecodeTables) -> &IndexVec<Self, Spanned<Self::Data>>;
+    fn index_vec_mut(mir: &mut FnBytecodeTables) -> &mut IndexVec<Self, Spanned<Self::Data>>;
 }
 
-impl LocalDecl {
-    pub fn new_return_place(return_ty: Ty) -> LocalDecl {
-        LocalDecl {
-            ty: return_ty,
-            name: None,
+pub trait MirIndexData: Sized + Clone {
+    type Index: MirIndex<Data = Self>;
+
+    fn index_vec(mir: &FnBytecodeTables) -> &IndexVec<Self::Index, Spanned<Self>> {
+        <<Self as MirIndexData>::Index as MirIndex>::index_vec(mir)
+    }
+
+    fn index_vec_mut(mir: &mut FnBytecodeTables) -> &mut IndexVec<Self::Index, Spanned<Self>> {
+        <<Self as MirIndexData>::Index as MirIndex>::index_vec_mut(mir)
+    }
+}
+
+impl AsRef<FnBytecodeTables> for FnBytecode {
+    fn as_ref(&self) -> &FnBytecodeTables {
+        &self.tables
+    }
+}
+
+impl AsRef<FnBytecodeTables> for Arc<FnBytecode> {
+    fn as_ref(&self) -> &FnBytecodeTables {
+        &self.tables
+    }
+}
+
+/// Permit indexing the HIR by any of the various index types.
+/// Returns the underlying data from the index, skipping over the
+/// span.
+impl<I> std::ops::Index<I> for FnBytecode
+where
+    I: MirIndex,
+{
+    type Output = I::Data;
+
+    fn index(&self, index: I) -> &I::Data {
+        &self.tables[index]
+    }
+}
+
+impl<I> std::ops::Index<I> for FnBytecodeTables
+where
+    I: MirIndex,
+{
+    type Output = I::Data;
+
+    fn index(&self, index: I) -> &I::Data {
+        &I::index_vec(self)[index]
+    }
+}
+
+indices::index_type! {
+    pub struct Place { .. }
+}
+
+#[derive(Copy, Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub enum PlaceData {
+    Variable(Variable),
+    Entity(Entity),
+    Field { owner: Place, name: Identifier },
+}
+
+#[derive(Copy, Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub enum LiteralData {
+    String(GlobalIdentifier),
+}
+
+indices::index_type! {
+    pub struct Variable { .. }
+}
+
+#[derive(Copy, Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub struct VariableData {
+    pub name: Identifier,
+}
+
+indices::index_type! {
+    pub struct Identifier { .. }
+}
+
+#[derive(Copy, Clone, Debug, DebugWith, PartialEq, Eq, Hash)]
+pub struct IdentifierData {
+    pub text: GlobalIdentifier,
+}
+
+/// Trait for the various types for which a span can be had --
+/// corresponds to all the index types plus `MetaIndex`.
+pub trait SpanIndex {
+    fn span_from(self, tables: &FnBytecodeTables) -> Span;
+}
+
+impl FnBytecode {
+    /// Get the span for the given part of the HIR.
+    pub fn span(&self, index: impl SpanIndex) -> Span {
+        index.span_from(&self.tables)
+    }
+}
+
+impl FnBytecodeTables {
+    /// Get the span for the given part of the HIR.
+    pub fn span(&self, index: impl SpanIndex) -> Span {
+        index.span_from(self)
+    }
+}
+
+impl<I: MirIndex> SpanIndex for I {
+    fn span_from(self, tables: &FnBytecodeTables) -> Span {
+        I::index_vec(tables)[self].span()
+    }
+}
+
+/// Declares impls for each kind of MIR index as well as the
+/// `mir::MetaIndex` enum.
+macro_rules! define_meta_index {
+    ($(($index_ty:ident, $data_ty:ty, $field:ident),)*) => {
+        $(
+            impl MirIndex for $index_ty {
+                type Data = $data_ty;
+
+                fn index_vec(mir: &FnBytecodeTables) -> &IndexVec<Self, Spanned<Self::Data>> {
+                    &mir.$field
+                }
+
+                fn index_vec_mut(
+                    mir: &mut FnBytecodeTables,
+                ) -> &mut IndexVec<Self, Spanned<Self::Data>> {
+                    &mut mir.$field
+                }
+            }
+
+            impl MirIndexData for $data_ty {
+                type Index = $index_ty;
+            }
+
+            debug::debug_fallback_impl!($index_ty);
+
+            impl From<$index_ty> for MetaIndex {
+                fn from(value: $index_ty) -> MetaIndex {
+                    MetaIndex::$index_ty(value)
+                }
+            }
+        )*
+
+        /// The HIR has a number of *kinds* of indices that
+        /// reach into it. This enum brings them together into
+        /// a sort of "meta index". It's useful sometimes.
+        #[derive(Copy, Clone, Debug, DebugWith, PartialEq, Eq, PartialOrd, Ord, Hash)]
+        pub enum MetaIndex {
+            $(
+                $index_ty($index_ty),
+            )*
+        }
+
+        impl SpanIndex for MetaIndex {
+            fn span_from(self, tables: &FnBytecodeTables) -> Span {
+                match self {
+                    $(
+                        MetaIndex::$index_ty(index) => index.span_from(tables),
+                    )*
+                }
+            }
+        }
+    };
+}
+
+define_meta_index! {
+    (Statement, StatementData, statements),
+    (BasicBlock, BasicBlockData, basic_blocks),
+    (Place, PlaceData, places),
+    (Variable, VariableData, variables),
+    (Identifier, IdentifierData, identifiers),
+    (Operand, OperandData, operands),
+    (Rvalue, RvalueData, rvalues),
+}
+
+/// A list of "MIR indices" of type `I`.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
+pub struct List<I: MirIndex> {
+    start_index: u32,
+    len: u32,
+    marker: std::marker::PhantomData<I>,
+}
+
+impl<I: MirIndex> Default for List<I> {
+    fn default() -> Self {
+        List {
+            start_index: 0,
+            len: 0,
+            marker: std::marker::PhantomData,
+        }
+    }
+}
+
+impl<I: MirIndex> List<I> {
+    /// Creates a list containing the values from in the
+    /// `start_index..end_index` from the enclosing `FnBodyTables`.
+    /// Ordinarily, you would not use this constructor, but rather
+    /// `from_iterator`.
+    fn from_start_and_end(start_index: usize, end_index: usize) -> Self {
+        assert_eq!((start_index as u32) as usize, start_index);
+        assert!(end_index >= start_index);
+
+        if start_index == end_index {
+            List::default()
+        } else {
+            List {
+                start_index: start_index as u32,
+                len: (end_index - start_index) as u32,
+                marker: std::marker::PhantomData,
+            }
         }
     }
 
-    pub fn new_temp(ty: Ty) -> LocalDecl {
-        LocalDecl { ty, name: None }
+    /// Creates a `List` containing the results of `from_iterator`.
+    pub fn from_iterator(
+        mut fn_bytecode: impl AsMut<FnBytecodeTables>,
+        iterator: impl IntoIterator<Item = I>,
+    ) -> Self {
+        let tables = fn_bytecode.as_mut();
+        let start_index = tables.list_entries.len();
+        tables
+            .list_entries
+            .extend(iterator.into_iter().map(|i| i.as_u32()));
+        let end_index = tables.list_entries.len();
+        List::from_start_and_end(start_index, end_index)
     }
 
-    pub fn new(ty: Ty, name: Option<String>) -> LocalDecl {
-        LocalDecl { ty, name }
+    pub fn is_empty(&self) -> bool {
+        self.len == 0
+    }
+
+    pub fn len(&self) -> usize {
+        self.len as usize
+    }
+
+    /// Iterate over the elements in the list.
+    pub fn iter(
+        &self,
+        fn_bytecode: &'f impl AsRef<FnBytecodeTables>,
+    ) -> impl Iterator<Item = I> + 'f {
+        let tables: &FnBytecodeTables = fn_bytecode.as_ref();
+        let start_index = self.start_index as usize;
+        let end_index = start_index + self.len as usize;
+        tables.list_entries[start_index..end_index]
+            .iter()
+            .cloned()
+            .map(I::from_u32)
+    }
+
+    /// Iterate over the data for each the element in the list.
+    pub fn iter_data(
+        &self,
+        fn_bytecode: &'f impl AsRef<FnBytecodeTables>,
+    ) -> impl Iterator<Item = I::Data> + 'f {
+        self.iter_enumerated_data(fn_bytecode).map(|(_, d)| d)
+    }
+
+    /// Iterate over the elements in the list *and* their associated
+    /// data.
+    pub fn iter_enumerated_data(
+        &self,
+        fn_bytecode: &'f impl AsRef<FnBytecodeTables>,
+    ) -> impl Iterator<Item = (I, I::Data)> + 'f {
+        let tables: &FnBytecodeTables = fn_bytecode.as_ref();
+        let data_vec = I::index_vec(tables);
+        self.iter(fn_bytecode).map(move |i| {
+            let data: &I::Data = &data_vec[i];
+            (i, data.clone())
+        })
     }
 }
 
-pub mod builtin_type {
-    #[allow(unused)]
-    pub const UNKNOWN: usize = 0;
-    pub const VOID: usize = 1;
-    pub const I32: usize = 2;
-    pub const STRING: usize = 3;
-    pub const ERROR: usize = 100;
-}
-
-#[derive(Debug)]
-pub enum Definition {
-    Builtin,
-    Fn(Function),
-    Struct(Struct),
-}
-
-pub struct Context {
-    pub definitions: Vec<Definition>,
-}
-
-impl Context {
-    pub fn new() -> Context {
-        let mut definitions = vec![];
-
-        for _ in 0..(builtin_type::ERROR + 1) {
-            definitions.push(Definition::Builtin); // UNKNOWN
-        }
-
-        Context { definitions }
-    }
-
-    pub fn add_definition(&mut self, def: Definition) -> usize {
-        self.definitions.push(def);
-        self.definitions.len() - 1
-    }
-
-    pub fn simple_type_for_def_id(&self, def_id: DefId) -> Ty {
-        Ty { def_id: def_id }
-    }
-
-    pub fn get_def_id_for_ty(&self, ty: Ty) -> Option<DefId> {
-        Some(ty.def_id)
-    }
-}
+debug::debug_fallback_impl!(for[I: MirIndex] List<I>);
