+++ components/lark-query-system/Cargo.toml
@@ -13,7 +13,7 @@ languageserver-types = "0.51.0"
 lark-entity = { path = "../lark-entity" }
 lark-error = { path = "../lark-error" }
 lark-hir = { path = "../lark-hir" }
-lark-mir2 = { path = "../lark-mir2" }
+lark-mir = { path = "../lark-mir" }
 lark-parser = { path = "../lark-parser" }
 lark-string = { path = "../lark-string" }
 lark-task-manager = { path = "../lark-task-manager" }
